backend/requirements.txt
openai>=1.0.0
jsonschema>=4.0.0
python-dotenv>=1.0.0
supabase>=2.0.0
psycopg2>=2.9.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0